+++ src/django_idom/__init__.py
@@ -1,7 +1,7 @@
-from . import hooks
+from . import components, hooks
 from .websocket.consumer import IdomWebsocket
 from .websocket.paths import IDOM_WEBSOCKET_PATH
 
 
-__version__ = "1.0.0"
-__all__ = ["IDOM_WEBSOCKET_PATH", "IdomWebsocket", "hooks"]
+__version__ = "1.1.0"
+__all__ = ["IDOM_WEBSOCKET_PATH", "IdomWebsocket", "hooks", "components"]