+++ tests/test_app/static/static-js-test.js
@@ -1,0 +1,3 @@
+let el = document.body.querySelector("#static-js");
+el.textContent = "StaticJS: True";
+el.dataset.success = "true";