+++ CHANGELOG.md
@@ -24,6 +24,16 @@ Types of changes are to be listed in this order
 
 -   Nothing (yet)
 
+## [1.1.0] - 2022-06-25
+
+### Added
+
+-   `django_css` and `django_js` components to defer loading CSS & JS files until needed.
+
+### Changed
+
+-   Bumped the minimum IDOM version to 0.39.0
+
 ## [1.0.0] - 2022-05-22
 
 ### Added
@@ -103,6 +113,7 @@ Types of changes are to be listed in this order
 -   Support for IDOM within the Django
 
 [unreleased]: https://github.com/idom-team/django-idom/compare/1.0.0...HEAD
+[1.1.0]: https://github.com/idom-team/django-idom/compare/1.0.0...1.1.0
 [1.0.0]: https://github.com/idom-team/django-idom/compare/0.0.5...1.0.0
 [0.0.5]: https://github.com/idom-team/django-idom/compare/0.0.4...0.0.5
 [0.0.4]: https://github.com/idom-team/django-idom/compare/0.0.3...0.0.4